refactor(ui): give the achievements and BIOS panes their own components
RomMGameInfoPanel held the state, the loads and the render code for all four tabs. The
achievements lane is fetched by the panel and read nowhere else in it, so it moves wholesale into
AchievementsTab: the three achievements fields, the lazy load with its load-once gate and offline
handling, and the badge rows.

BiosTab takes the BIOS render code and its builders, and nothing else. biosStatus, biosLevel and
coreInfo stay in PanelState, because a core change has to reach the render as ONE update — two
would briefly highlight the previous core and name it in the "Active Core" row against the new
core's requirements. The core-info fetch and handleCoreChange's single write are untouched.
PanelState ends at 24 fields, the panel at 1078 lines from 1623.

Both tabs mount for every ROM and render nothing until their tab is active. For achievements that
is load-bearing: unmounting on a tab switch would re-fetch and re-spinner on every visit. Its key
is the rom id, so a version switch remounts it, and that remount is how its per-rom state and its
load-once gate re-key.

The load counter the two lazy lanes shared through a panel ref moves to connectionState, next to
the retry-progress store it guards: beginServerLoad hands out a claim, settleServerLoad clears the
store only while that claim is still the newest one taken out.

infoRow and section move to panelSection so the panel and the BIOS pane share one definition
instead of a copy each.

Two behaviors differ, both deliberately. The load counter is module-global where the ref was per
panel instance, so two mounted panels can no longer wipe each other's retry frames. The other is a
fix that falls out of the re-key: the version-switch handler used to clear the achievements list
unconditionally, so an event that did not change the rom id wiped the list without re-loading it,
and the tab read "No achievements found for this game" until it was left and re-entered. Keyed by
rom id, that same event now leaves the correct list standing.

The rom-binding invariant register described a panel with two lazy lanes. It has one, and the
achievements tab is named there as bound by its own React key.

RomMGameInfoPanel.test.tsx is unchanged. Its 130 tests pass against the new structure as they did
against the old, which is what makes this a move rather than a rewrite.

docs: N/A — pure refactor; both panes render exactly what they rendered.

Part of #994; the state split and the JSX conversion follow separately.
